+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 0.6.1 - January 2015
+### Changed
+- Setting config options in block style will last for all following `convert` calls.
+- Inline config options are only applied to this particular operation
+
+### Removed
+- `config.reset` is removed
 
 ## 0.6.0 - September 2014
 ### Added